Fix parsing of identifiers in type contexts
This fixes a historical bug the `wast` crate when parsing `$id`-style
identifiers in exports within a type context (e.g. within an instance
type or a component type). This was due to what I believe was a
historical misunderstanding of mine and/or maybe some historical change,
unsure!

This is a breaking change for the text format but given how rarely this
should be used it's expected that this breaking change practically won't
actually have any impact.

Closes #2192