Consolidate caches used when encoding WIT types
This commit fixes an issue with the previous commit by introducing a new
cache of `&TypeDefKind` to encoded type during the WIT encoding process.
This addition of a new cache, in addition to the two previous encoding
caches, necessitated a refactoring of wrapping up these caches into a
single type.

This change means that if the same WIT type shows up as multiple
`TypeId` instances it'll only be encoded as a single type in the final
wasm binary. This ensures that the text format of WIT matches the binary
format of WIT and packages all round-trip correctly, especially in the
face of `include`.