Revise MSRV policy for this repository
In writing the current policy for this repository's minimum supported
version of Rust (MSRV) is "stable minus two" which matches Wasmtime.
This does not reflect the current reality, however, where the current
MSRV is 1.76 where the current Rust is 1.87. This additionally does not
reflect what I think is the best decision for this repository. While
"stable minus two" matches Wasmtime this repository is expected to be
possibly used in more contexts than Wasmtime, for example any
wasm-processing tools which can be lower-level than anything needing a
wasm runtime.

Given all this the change here is to update the comments on MSRV policy
for this repository to "stable minus 10, not proactively updated". That
is a much larger window of Rust but the code in this repository is much
less likely than Wasmtime, for example, to benefit from newer features
of Rust. The hope is that this won't be too onerous to maintain while
still allowing to eventually use newer features as necessary.