Turn on new WIT metadata format by default
This commit is a follow-up to #1508 where a new JSON format was added
for storing docs/stability and accounting for world imports/exports
allowing overlap. This support should have permeated far enough that
it's reasonable to turn this on-by-default now.

This additionally fixes fuzz bugs cropping up after adding support for
stability annotations in the generated WIT, the other motivation for this.