Simplify arity implementation (#2236)
Originally the annotations in the `for_each_operator!` macro were
intended to be used to implement the validator in addition to arity, but
during the PR this part was removed and the annotations were only used
for arity. As such the macro/annotations were a bit more complicated
than they might otherwise need to be. Coupled with #2235 where some
changes may be needed to arity calculations the macro was a bit
difficult to understand. As a result this commit simplifies the arity
annotations to either:

* `arity N -> N`
* `arity custom`

All "custom" instructions are now written as Rust functions in the
`arity.rs` module. They delegate to each other internally as appropriate
in a similar manner to how the macro calculated before.

Overall this passes tests but I can clarify that I only have a tenuous
understanding at best of the previous macro. I did my best to faithfully
translate it from what it used to be but it's possible I got something
wrong along the way.