Extract `json-from-wast` to its own crate (#2247)
* Extract `json-from-wast` subcommand to its own crate

Polish it up a bit for publication to crates.io to get used externally
as well, for example in Rust projects that want to consume the AST
produced here.

* Refine the AST of json-from-wast

This mostly doesn't change the JSON representation or how things are
serialized, but this does change the Rust-level representation to make
it easier to consume from a Rust crate (less stringly-typed). A few
minor adjustments were made and additionally new structures are
supported (e.g. component model types)

* Undo some `dbg!`

* Undo some `gc.rs` changes

* Fix MSRV compat

* Add to publish list

* Remove readme directive