Don't fail to print known custom sections (#2264)
This commit updates the `wasm-tools print` subcommand to print raw
custom sections for some "known" custom sections if there is not printer
registered for it. For example the `linking` and `reloc.CODE` custom
sections are "known" but have no text format. Previously they would fail
to print but now they will print the raw bytes.