fail validation if callback missing for async export
If the module exports a function with the `[async-lift]` prefix (rather than the
`[async-lift-stackful]` one), then we must be able to match that up with a
corresponding callback function; otherwise, we fail validation.

Prior to this fix, we'd silently generate an invalid component and leave it to
`wasmparser` to notice, which is bad form, plus the diagnostic message wasn't
very helpful.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>