Add support for tags to `wasm-tools component link` (#2306)
* Add support for tags to `wasm-tools component link`

While not explicitly documented it looks like dynamic libraries all
import exception tags and expect the dynamic linker to provide the same
tag item to all modules importing the same name. This updates creation
of the `env` module to contain tags in addition to the other items it
already contains.

* Fix numbering of indirect table entries added