Print offset of final `end` in `wasm-tools print -p` (#2312)
I was looking at a validation error recently where the offset printed
for an error wasn't present in `wasm-tools print -p`, and that was
because the error was on the final `end` instruction of a function
which isn't printed in the text format. This commit updates the output
of `-p` to include a comment with an offset for the `end` instruction.