feat(compose): enable all features during component validation
This commit enables all wasm features (i.e. `WasmFeatures`) for the
`wasm-tools compose` subcommand, so newer components with enabled
features (e.g. p3 async components) can be composed.

While wasm-tools compose is deprecated in favor of wac, it is
currently the only composition tool of the two that supports p3.