add `CloneMaps` type and parameter to `Resolve::merge_worlds`
This allows the caller to determine which types and/or interfaces were cloned as
part of the merge operation.

Note that I made this an `&mut` parameter to `Resolve::merge_worlds` rather than
a return value in order to make it a bit more efficient when calling
`Resolve::merge_worlds` multiple times and collecting the clones across all
those calls.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>