add `[async-lower]` prefix to `read` and `write` imports
I realized that some of my `componentize-py` tests weren't actually testing what
they were supposed to, and that was due to the `{stream|future}.{read|write}`
intrinsics being lowered synchronously :facepalm:.  This fixes that.

Note that we're still lowering the `{stream|future}.cancel-{read|write}`
intrinsics synchronously, same as the Rust `wit-bindgen` does.  We might need to
change that (or make it configurable) for toolchains that actually have a good
async cancellation paradigm.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>