Fix parsing named locals without an inline function type
This commit fixes a bug in `wast` where the generated `name` section was
invalid and placed local names on the wrong location. This happened when
a function type was not listed inline for a function (e.g. it was `(type
$foo)`) meaning that the binary translation started with the wrong local
offset. The fix here is to do a lookup of the function type to see how
many parameters it has and start the indexing there, or skip emitting
local names if the function type doesn't exist.