Support special import names for unit futures/streams (#2390)
* Support special import names for unit futures/streams

This commit enhances `wit-component`'s import detection to support
"unit" `future` and `stream` types which don't have a payload in a
special way. Previously it was required that these types were mentioned
in some WIT somewhere to get connected to a function, but now it's
possible to import these intrinsics directly even if the type isn't
mentioned in WIT.

The motivation for this change is to support guest language async
runtimes which are world-agnostic to still operate on unit
futures/streams as necessary without binding them to any particular WIT
import or export. For example it's expected that unit streams are a
primitive for inter-task communication in the component model.

* Update crates/wit-component/src/validation.rs

Co-authored-by: Joel Dice <joel.dice@fermyon.com>

---------

Co-authored-by: Joel Dice <joel.dice@fermyon.com>