Fix foreign interface dependencies which are disabled by feature-gating
A foreign imported interface can be missing from pkg.interfaces either
because it does not exist, or because it was disabled by feature-gating.

Previously, the parser failed in both cases, when it should only
fail when it does not exist, and should succeed when the references
to it are disabled by feature-gating.

By checking the set of unresolved worlds' imports and exports for
references to this interface, it is possible to determine if they
contain any references to the interface, and whether or not they
were disabled by feature gating.

If they are all disabled, then it does not matter that the interface
is absent from pkg.interfaces, and the code should not fail. If any
references are active however, then this is an error as the code
will not be able to find the interface in pkg.interfaces.

If no references exist at all, then no combination of feature
enabling/disabling would satisfy the import/use statement which
requires the interface, and the code should fail.

This fixes issue 2285 and a unit test was added to cover this case.