wit-parser: Use stability info when processing foreign deps
Previously, the process_foreign_interfaces function would
receive unresolved interfaces with their stability values
set to Unknown

This made it hard to distinguish between interfaces which
had been feature gated out, and genuinely nonexistent
interfaces

This commit records the stability of each import of a foreign dependency
when populating the self.foreign_{deps, interfaces, worlds}
maps.

This enables process_foreign_interfaces to push None as a placeholder
if it sees that all of its imports have been feature-gated out. Otherwise,
it will allow the resolver to try and resolve the import, and if the
interface genuinely does not exist, it will fail.

Fixes issue 2285